backend/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
facenet-pytorch>=2.5.2
opencv-python>=4.5.3
grad-cam>=1.3.0
Pillow==10.4.0
mss>=6.1.0
numpy
albumentations>=1.3.0
scikit-learn>=0.24.0
tqdm>=4.62.0
pyyaml>=5.4.0
flask>=2.0.0
flask-cors>=3.0.10